Refactor `test_strategy_templates.py` into tests per module
