+++ tests/strategies/test_mean_reversion.py
@@ -1,0 +1,24 @@
+import polars as pl
+from quant_trading_strategy_backtester.strategies.mean_reversion import (
+    MeanReversionStrategy,
+)
+
+
+def test_mean_reversion_strategy_initialization() -> None:
+    params = {"window": 20, "std_dev": 2.0}
+    strategy = MeanReversionStrategy(params)
+    assert strategy.window == 20
+    assert strategy.std_dev == 2.0
+
+
+def test_mean_reversion_strategy_generate_signals(
+    mock_polars_data: pl.DataFrame,
+) -> None:
+    params = {"window": 5, "std_dev": 2.0}
+    strategy = MeanReversionStrategy(params)
+    signals = strategy.generate_signals(mock_polars_data)
+    assert isinstance(signals, pl.DataFrame)
+    EXPECTED_COLS = {"signal", "mean", "std", "upper_band", "lower_band", "positions"}
+    for col in EXPECTED_COLS:
+        assert col in signals.columns
+    assert signals["signal"].is_in([0.0, 1.0, -1.0]).all()
